Initial commit, moved GameLogicTests from PlayerTest to GameLogicTests, playerAlive status to Player.
