+++ src/game/Gameboard.java
@@ -13,13 +13,11 @@ public class Gameboard {
     private Levels levels;
     private Player player;
 
-    private boolean isAlive = true;
     private boolean isFinished = false;
 
     public Gameboard() {
         levels = new Levels(level);
-        player = new Player();
-        player.setName("Kalle");
+        player = new Player(9,1, CHARACTER);
         boardGrid = levels.getBoard();
         boardGrid[characterPosition.getY()][characterPosition.getX()].setTileType(CHARACTER);
         boardGrid[doorPosition.getY()][doorPosition.getX()].setTileType(DOOR);
@@ -27,17 +25,13 @@ public class Gameboard {
         boardGrid[10][3].setTileType(MONSTER);
     }
 
-    public boolean getPlayerAlive() {
-        return isAlive;
-    }
-
-    public void setPlayerAlive(boolean playerAlive) {
-        this.isAlive = playerAlive;
-    }
-
     // Returning the Point of requested position
     public Point getPoint(int y, int x) {
         return boardGrid[y][x];
+    }
+    
+    public Player getPlayer() {
+    	return player;
     }
 
     // Moving character in desired direction
@@ -74,7 +68,6 @@ public class Gameboard {
         } else if (p.getTileType() == MONSTER) {
             // Returns true to show that character really stepped onto monster
             // Game is although over
-            isAlive = false;
             return false;
         } else if (p.getTileType() == TREASURE) {
             player.setTreasure(1);
@@ -88,10 +81,6 @@ public class Gameboard {
             // Otherwise movement is a okay
             return false;
         }
-    }
-
-    public boolean getIsAlive() {
-        return isAlive;
     }
 
     public boolean isFinished() {
