+++ src/game/Player.java
@@ -1,14 +1,18 @@
 package game;
 
 /**
- * @author Marcus
+ * @author Marcus Laitala
  * @Date 2019-03-17
  * @version 0.3
  */
-public class Player {
+public class Player extends Point{
 
-	private String name;
+	public Player(int y, int x, TileType tileType) {
+		super(y, x, tileType);		
+	}
+	
 	private int treasures;
+    private boolean isAlive = true;
 	
 	public int getTreasure() {
 		return treasures;
@@ -18,11 +22,7 @@ public class Player {
 		this.treasures = treasures;
 	}
 
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}	
+	public boolean getAlive() {
+        return isAlive;
+    }
 }
